12.0 session preserve forward port pc

+++ session_redis/http.py
@@ -76,6 +76,20 @@ def purge_fs_sessions(path):
             pass
 
 
+def copy_fs_sessions(path):
+    from odoo.http import OpenERPSession
+    from werkzeug.contrib.sessions import FilesystemSessionStore
+    werkzeug_session_store = FilesystemSessionStore(path, session_class=OpenERPSession)
+    session_store = http.Root().session_store
+    filename_prefix_len = len('werkzeug_')
+    filename_suffix_len = len('.sess')
+
+    for fname in os.listdir(path):
+        session_file = fname[filename_prefix_len:filename_suffix_len * -1]
+        session = werkzeug_session_store.get(session_file)
+        session_store.save(session)
+
+
 if is_true(os.environ.get('ODOO_SESSION_REDIS')):
     if sentinel_host:
         _logger.debug("HTTP sessions stored in Redis with prefix '%s'. "
@@ -87,5 +101,9 @@ if is_true(os.environ.get('ODOO_SESSION_REDIS')):
 
     http.Root.session_store = session_store
     http.session_gc = session_gc
-    # clean the existing sessions on the file system
-    purge_fs_sessions(odoo.tools.config.session_dir)
+
+    if is_true(os.environ.get('ODOO_SESSION_REDIS_COPY_EXISTING_FS_SESSIONS')):
+        copy_fs_sessions(odoo.tools.config.session_dir)
+    if is_true(os.environ.get('ODOO_SESSION_REDIS_PURGE_EXISTING_FS_SESSIONS')):
+        # clean the existing sessions on the file system
+        purge_fs_sessions(odoo.tools.config.session_dir)
